Add support for API Tokens
This commit adds provisional support for accepting and disambiguating between API keys and API
Tokens, provided in the same field of the underlying data store. Depending on the format of the
value given, we set appropriate authentication headers, to be accepted in API calls to Cloudflare.

The target of the test request has also been updated from `/user` to `/accounts`, which is available
to both API keys and API Tokens.