sanitise sensitive HTTP headers in logAPICall
There have been previous attempts at preventing sensitive information from
appearing in the logs (see #45) however that fix only accounted for when the
logging was called as a result of `callAPI`. For whatever reason, there are
also uses whereby the `logAPICall` is used as a general purpose logger with
portions of the HTTP request included which were skipping the sanitisation in
`callAPI` and ended up in user logs.

In order to address the direct calls, I've moved the sanitisation logic from the
HTTP request to the Logger. This means instead of attempting to play
whack-a-mole with new usages of where we could potentially log, it is redacted
at the point of writing out the log entry which is easier to use as a
chokepoint due to it's central use.

Another choice I've made as a part of this change is that I've also swapped from
removing the HTTP header to just redacting the value. The reason for this is
that while debugging or preparing logs, it is useful to be able to pin point
the authentication mechanism in use as that also helps determines the potential
causes of the failures.

Following this fix, the log output now shows REDACTED instead of the actual
value.

```
[Tue Mar 16 23:17:51.144185 2021] [php7:notice] [pid 86] [client 172.25.0.1:65128] [Cloudflare] ERROR: [CLIENT API] Array
(
    [type] => request
    [path] => https://api.cloudflare.com/client/v4/zones?action=cloudflare_proxy
    [timeout] => 30
    [method] => GET
    [headers] => Array
        (
            [Content-Type] => application/json
            [X-Auth-Email] => REDACTED
            [X-Auth-Key] => REDACTED
        )

        )
 , referer: http://wordpress.lol:8000/wp-admin/options-general.php?page=cloudflare
```

Closes cloudflare/Cloudflare-WordPress#389